requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
azure-storage-queue==12.11.0
azure-storage-blob==12.22.0
